restructuring calendar component
